+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3tuning 0.17.1.9000
+# mlr3tuning 0.17.2
 
 * feat: `AutoTuner` accepts instantiated resamplings now.
   The `AutoTuner` checks if all row ids of the inner resampling are present in the outer resampling train set when nested resampling is performed.